+++ ai_ui/forms/postcode.py
@@ -1,0 +1,11 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, RadioField, SubmitField, validators
+
+
+class PostcodeForm(FlaskForm):
+    postcode = StringField('Postcode', [validators.DataRequired("Please enter a postcode.")])
+    classificationFilter = StringField('Classification Filter')
+    historical = RadioField('Include historical data?', [validators.DataRequired()],
+                            choices=[('true', 'Yes'), ('false', 'No')], default='true')
+    resultsPerPage = StringField('Results per page', default='10')
+    submit = SubmitField('Search')
